Implemented CRI and clenaup callback.
CRI
===

CRI (standing for "Common Resources Interface") is simply a shared
object that is passed to all method implementations as third parameter.

It allows all API method implementations to share data, functions or any
kind of resources even when spreaded across multiple files withot
needing to separately require them as modules in each file.

It can be initialyzed by passing it as third parameter to PASAR
constructor. But, if omitted, it automatically defaults to empty object.
So you doesn't need to define it explicitly at all. It will even be
available for data sharing to all method implementations which declare
it as (third) parameter.

I called it "cri" to avoid calling it "res", standing for "resources" to
avoid confusion with the Express response object (even we thankfully
sould almost never need to deal with it).

Cleanup callback
================

Appart of "cri", I added new fourth parameter that will be always
injected to method implementations.

I called it "pbk" standing for "Promise-callback".

It is intended to be a placeholder to return data or resources back from
method implementation (even inside promise implementation).

First use of such thing if for the "cleanup callback" which, if provided
(and is a function) will be called in case of timeout (if specified)
happen.

WARNING: You should be carefull to define it BEFORE any chance returning
promise could be resolved neither with fullfillment or rejetion
(including timeout). That is: NOT in any async operation.