Improved form facility to allow file uploads.
File uploads requires multipart POSTs so server needs express-fileupload
(or compatible) middleware to be used.

This is achieved by using FormData object in the Ajax form submission
but, to avoid to always require multipart support server side, the
presence of (actually filled) file inputs is detected at submission time
and FormData is used ONLY if it is strictly requierd.